+++ execution/src/main/java/com/multiassetoms/execution/model/OrderStatus.java
@@ -2,5 +2,7 @@ package com.multiassetoms.execution.model;
 
 public enum OrderStatus {
     CREATED,
-    SENT
+    SENT,
+    ACKED,
+    REJECTED
 }